General fixes (#68)
* Swapped values for it to calculate correctly
* Uncommented tests
* Enabled more tests
* Removed comments
* Global and local redfine tests working (Disabled test for windows since manual works but not via tests)
* Disabled debug in symbol table

+++ tests/testfiles/scope/local.m
@@ -4,6 +4,6 @@ setup()
 
 mainloop()
 {
-    int8 test = 100;
+    int16 test = 100;
     print(test);
 }
